requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djoser>=2.1,<2.2
gunicorn>=20.1,<21.0
psycopg2-binary>=2.9,<3.0
